+++ src/PaooGame/Creator/ItemCreator.java
@@ -6,10 +6,10 @@ import PaooGame.RefLinks;
 
 public abstract class ItemCreator {
 
-    public Item getItem(ItemType type, RefLinks refLink, float x, float y) {
+    public Item getItem(ItemType type, RefLinks refLink, int x, int y) {
         Item item = createItem(type, refLink, x, y);
         return item;
     }
 
-    protected abstract Item createItem(ItemType type, RefLinks refLin, float x, float y);
+    protected abstract Item createItem(ItemType type, RefLinks refLin, int x, int y);
 }
