+++ src/PaooGame/Items/Ogre.java
@@ -6,7 +6,7 @@ import java.awt.*;
 
 public class Ogre extends Character {
 
-    public Ogre(RefLinks refLink, float x, float y)
+    public Ogre(RefLinks refLink, int x, int y)
     {
         super(refLink, x, y, Character.DEFAULT_CREATURE_WIDTH, Character.DEFAULT_CREATURE_HEIGHT);
     }
